Always import cc, even if it's useless
This is required for `build.rs`, even if it returns right away
because the pure rust versions will be compiled.

Ideally, we should not import cc at least if `pure-rust` is defined,
but `cargo` doesn't seem to have any ways to do that, and `#[cfg]`
macros wouldn't work in `build.rs` anyway.